\name{citest-package}
\alias{citest-package}
\alias{citest}
\docType{package}
\title{
\packageTitle{citest}
}
\description{
\packageDescription{citest}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{citest}

}
\author{
\packageAuthor{citest}

Maintainer: \packageMaintainer{citest}
}
\keyword{ package }
\seealso{\code{\link{foo}}}
